requirements.txt
Flask==2.1.2
flaskwebgui==0.3.5
psutil==5.9.0
PyYAML==6.0
requests==2.27.1